# Improve performance of `CommutationAnalysis` transpiler pass