@ewinston 2021-09-21
I checked scaling of timing and memory usage of QV circuits and didn't see any major issues. LGTM.

peak memory usage vs square QV circuit.
![image](https://user-images.githubusercontent.com/25756217/134212724-76f9ce51-03cc-4e6a-aa3d-47379fb06d05.png)

total time spent in function
![image](https://user-images.githubusercontent.com/25756217/134212879-84835cec-8526-456f-b7b9-1f7d6fca130c.png)